+++ frontend-ios/YakAl/YakAl/Sources/Views/Calendar/CalendarSwiftUIView.swift
@@ -1,107 +1,33 @@
 import SwiftUI
 
-//MARK: - Calendar Week View : 매 주차 뷰
-struct CalendarWeekView: View {
-    var weekDates: [DayModel]
-    @Binding var selectedDate: DayModel? // 선택한 날짜를 바인딩
-
-    var body: some View {
-        HStack(spacing:20) {
-                   ForEach(weekDates, id: \.self) { date in
-                       let isSelectedDate = date.dayText == selectedDate?.dayText && date.date.month == selectedDate?.date.month
-
-                       if isFirstWeek() && date.date.day == "1" {
-                           if isLastOne(in: weekDates, current: date) {
-                               CircularProgressBar(progress: 0.9, dateText: date.dayText, today: isSameDate(date.date, Date()), weekDates: date.weekdayText, isSelected: isSelectedDate)
-                                   .padding(.top, 4)
-                                   .onTapGesture {
-                                       selectedDate = date
-                                   }
-                           } else {
-                               CircularProgressBar(progress: 0, dateText: "  ", today: false, weekDates: date.weekdayText, isSelected: false)
-                                   .padding(.top, 4)
-                           }
-                       } else if isLastWeek() && (Int(date.date.day)! <= 7 || date.date.day == "1") {
-                           CircularProgressBar(progress: 0, dateText: "  ", today: false, weekDates: date.weekdayText, isSelected: false)
-                               .padding(.top, 4)
-                               .onTapGesture {
-                                   selectedDate = date
-                               }
-                       } else {
-                           CircularProgressBar(progress: 0.2, dateText: date.dayText, today: isSameDate(date.date, Date()), weekDates: date.weekdayText, isSelected: isSelectedDate)
-                               .padding(.top, 4)
-                               .onTapGesture {
-                                   selectedDate = date
-                               }
-                       }
-                   }
-               }.padding(.horizontal,20)
-        
-      }
-    func isSameDate(_ date1: Date, _ date2: Date) -> Bool {
-        let calendar = Calendar.current
-        return calendar.isDate(date1, inSameDayAs: date2)
-    }
-    
-    func isFirstWeek() -> Bool {
-        return weekDates.contains(where: { $0.date.day == "1" })
-    }
-    
-    
-    func isLastOne(in dates: [DayModel], current: DayModel) -> Bool {
-        if let lastOne = dates.last(where: { $0.date.day == "1" }) {
-            return lastOne.id == current.id
-        }
-        return false
-    }
-    
-    func isLastWeek() -> Bool {
-        return weekDates.contains(where: { $0.date.day == "28" || $0.date.day == "29" || $0.date.day == "30" || $0.date.day == "31" })
-    }
-    
-    
-    
-    
-    func weekdayColor(for weekdayText: String) -> Color {
-        switch weekdayText {
-        case "일":
-            return Color(red: 0.88, green: 0.06, blue: 0.16)
-        case "토":
-            return Color(red: 0.33, green: 0.53, blue: 0.99)
-        default:
-            return Color.black
-        }
-    }
-}
 //MARK: - CalendarSwiftUIView : 달력 뷰
 struct CalendarSwiftUIView: View {
-    // --------------- 상태 ---------------
+    
+    /* --------------- State --------------- */
+    //currentDate 현재 일자를 불러온다.
     @State private var currentDate = Date()
+    //isExpanded 확장 여부
     @State private var isExpanded = false
-    
+    // currentWeekIndex 현재 주
     @State private var currentWeekIndex: Int = 0
-    
     // 선택한 날짜를 저장할 상태 변수 추가
     @State private var selectedDate: DayModel? = DayModel(currentDate: Date())
-
-
-    /* --------------- 요일 스트링 --------------- */
     
-
-    // 년 월 표현
+    /* --------------- dateFormatter 년 월 포멧  --------------- */
     var dateFormatter: DateFormatter {
         let formatter = DateFormatter()
         formatter.dateFormat = "yyyy년 M월"
         return formatter
     }
     
-    // 예시 약물
+    /* --------------- 예시 약물 --------------- */
     var sampleMedications: [Medicine] = [
         Medicine(id: 1, image: "image_덱시로펜정", name: "약물A", ingredients: "성분 A", dangerStat: 0, isTaken: false),
         Medicine(id: 2, image: "image_덱시로펜정", name: "약물B", ingredients: "성분 B", dangerStat: 1, isTaken: false)
         
     ]
     
+    /* --------------- 현재 주 찾기 --------------- */
     func findCurrentWeekIndex() -> Int {
         let today = Date()
         for (index, week) in sampleWeekDates.enumerated() {
@@ -114,10 +40,10 @@ struct CalendarSwiftUIView: View {
         return 0 // Default to the first week if not found
     }
     
+    
     /* --------------- 한줄 캘린더 데이터--------------- */
     var sampleWeekDates: [[DayModel]] {
         var allWeekDates: [[DayModel]] = []
-        
         let calendar = Calendar.current
         let components = calendar.dateComponents([.year, .month], from: currentDate)
         guard let startOfMonth = calendar.date(from: components),
@@ -146,11 +72,6 @@ struct CalendarSwiftUIView: View {
         
         return allWeekDates
     }
-
-
-
-    
-    
     /* --------------- 토,일 색상 ---------------*/
     func weekdayColor(for weekdayText: String) -> Color {
         switch weekdayText {
@@ -193,7 +114,7 @@ struct CalendarSwiftUIView: View {
             }.padding(.top, 20)
                 .padding(.bottom, 15)
             
-            // --------------- 요일  ---------------
+            // --------------- 요일 불러오는 Header  ---------------
             WeekdayHeaderSwiftUIView()
             
 
@@ -203,11 +124,11 @@ struct CalendarSwiftUIView: View {
                 // 매일자 삽입
                 ForEach(sampleWeekDates, id: \.self) {
                     weekDates in
-                    CalendarWeekView(weekDates: weekDates, selectedDate: $selectedDate) // 선택한 날짜를 하위 뷰에 전달
+                    CalendarWeekSwiftUIView(weekDates: weekDates, selectedDate: $selectedDate) // 선택한 날짜를 하위 뷰에 전달
                 }
             } else {
                 // 해당 주차
-                CalendarWeekView(weekDates: sampleWeekDates[currentWeekIndex], selectedDate: $selectedDate) // 선택한 날짜를 하위 뷰에 전달
+                CalendarWeekSwiftUIView(weekDates: sampleWeekDates[currentWeekIndex], selectedDate: $selectedDate) // 선택한 날짜를 하위 뷰에 전달
 
                 
             }
